Re-implement graphics primitives in Idris
Right now this just makes rendering dramatically slower. (I think
the bottleneck is just writing data to the texture one pixel at a
time, which there's not a whole lot I can do about). It is in
preparation for making the graphics better though. I'll be able
to add antialiasing and higher line weights and stuff.