Start implementing stroke thickness (incomplete code)
This code doesn't actually work in its current state, but I realised
I should take it in a different direction, and wanted to save the
current state just in case. It currently separates the polygon to
draw into vertical sequences of line segments, which is unnecessary.
I've implemented drawing for circles and dots kind of completely
(there are still improvements to do), added a new primitive of a
sequence of connected beziers (with an arbitrary number of control
points each), implemented generation of outlines for beziers, and
not yet finished implementing the algorithm to fill those outlines.