Avoid drawing offscreen better
Only loop over those pixels that are actually in view when drawing
circles and beziers. Cache the convex hulls of composite pictures
so they aren't re-calculated every time. Fixed a bug in hull calculation
by replacing it with a completely different more numerically stable
algorithm (which is probably faster too because it doesn't use atan2,
but is technically quadratic in the worse case like quicksort).