Rewrite the slow bits of the renderer in C
It turns out making thousands of calls to edit pointers in Idris is
super slow, so now it converts the beziers to polygons in Idris then
sends the polygons to C to turn into pixels. VectorToPixel.idr is now
very mis-named. I should do something about that at some point.